requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
datasets>=2.20
tqdm>=4.66
pyyaml>=6.0
numpy>=1.26
